+++ cardinal_pythonlib/run_all_tests.py
@@ -1,5 +1,5 @@
 #!/usr/bin/env python
-# cardinal_pythonlib/version_string.py
+# cardinal_pythonlib/run_all_tests.py
 
 """
 ===============================================================================
@@ -22,14 +22,23 @@
 
 ===============================================================================
 
-**Current version number of this library.**
-
-NOTE: this file must be importable by setup.py during package installation and
-must therefore have NO DEPENDENCIES (e.g. semantic_version).
-
-For changelog, see changelog.rst
+**Run all tests across the library.**
 
 """
 
-VERSION_STRING = '1.1.6'
-# Use semantic versioning: http://semver.org/
+import subprocess
+import os
+
+
+THIS_DIR = os.path.dirname(os.path.realpath(__file__))
+
+
+def run_all_unit_tests() -> None:
+    cmdargs = [
+        "python", "-m", "unittest", "discover", "-s", THIS_DIR, "-p", "*.py"
+    ]
+    subprocess.check_output(cmdargs)
+
+
+if __name__ == "__main__":
+    run_all_unit_tests()